Methods Ex1
Write a program that accepts two numbers from the user via a
JOptionPane. Your program must be defined with four methods. Each of
these methods accepts two integer values and returns an integer. Each of
these methods is to be called in turn from the main method passing to
them the values accepted from the user). The main method must capture
and output the return values from each of these methods.
• Method1 – Accepts 2 integer values, adds the values together & returns
the answer.
• Method2 – Accepts 2 integer values, subtracts the first value from the
second & returns the answer
• Method3 – Accepts 2 integer values, multiples the first value by the
second & returns the answer to where it was called.
• Method4 – Accepts two integer values, divides the first number by the
second and returns the answer to where it was called.