Methods Ex 4
Write a program that prompts a teacher to enter 10 marks (as ints) that
students in their class have achieved in a recent class test.
• These marks must be stored in an array. The array is then to be passed
to a method which will output the grade which is applicable to each
mark.
• The grading system is as follows:
• A mark less than 40 is a fail.
• A mark greater than or equal to 40 but less than 50 is a pass.
• A mark greater than or equal to 50 but less than 60 is a merit.
• A mark greater than or equal to 60 but less than or equal to 100 is a
distinction.
• If the mark is less than 0 but greater than 100 then the method should
print out a message indicating that the mark is outside the allowable
range.